kirkwood:  Migrate uci config to new PCIe path
The name of the PCIe controller node in device tree changed
between kernel 4.7 and kernel 4.14. Migrate the configuration
when an update from kernel 4.7 to 4.14 is done to make the
existing wifi configuration compatible with the new names.

This commit for kirkwood is largely based on commit 0bd5aa89fcf
for mvebu.  (Commit b452af23 also performs a migration for
mpc85xx that is somewhat similar to this.)

The need for this was found when upgrading from 17.01.4 to
18.06.2.  This post also identifies the issue:
https://forum.openwrt.org/t/wireless-not-functional-after-18-06-upgrade-extra-radios-detected/18065

Signed-off-by: J Steele joe@madewell.net